Basic Account functionality update
Adds a Member class object representing the user's Petlibro account, and makes modifications to accommodate it.
Allows the user to update some of their account settings, as well as show or use those account settings on the front-end with a sensor entity (lovelace, automations, scripts etc)
Can be easily built upon to include more settings and ways to manage things on your account in the future.